Mapping moved to parser utils
